+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Compatibility:
 * Implemented `rb_uv_to_utf8` (#1998).
 * Implemented `rb_str_cat_cstr`.
 * Implemented `rb_fstring`.
+* Implemented `rb_enc_str_buf_cat`.
 
 Performance: